requirements.txt
# 齐齐哈尔大学教务系统查询工具 v2.0.0
# Python依赖包列表 - 所有版本已固定以确保构建的可重现性
# 更新日期: 2025-07-17

# ==================== 核心框架 ====================
kivy==2.3.1                    # 跨平台GUI框架
pillow==10.4.0                 # 图像处理库，支持多种图像格式

# ==================== 网络和HTTP ====================
requests==2.32.3               # 简洁优雅的HTTP库
urllib3==2.2.2                 # 强大的HTTP客户端库
certifi==2024.7.4              # Mozilla CA证书包

# ==================== HTML/XML解析 ====================
beautifulsoup4==4.12.3         # 强大的HTML/XML解析库
lxml==5.2.2                    # 高性能XML/HTML解析器

# ==================== 字符编码处理 ====================
chardet==5.2.0                 # 字符编码自动检测
charset-normalizer==3.3.2      # 字符集规范化库
idna==3.7                      # 国际化域名支持

# ==================== 开发和调试工具（可选） ====================
# 取消注释以下行来启用开发工具
# pytest==8.2.2                # 现代化单元测试框架
# black==24.4.2                 # Python代码格式化工具
# flake8==7.1.0                 # 代码质量检查工具
# mypy==1.10.1                  # 静态类型检查器

# ==================== 构建工具（仅用于APK构建） ====================
# 这些依赖仅在使用buildozer构建Android APK时需要
# buildozer==1.5.0             # Android APK构建工具
# cython==3.0.10               # Python到C的编译器